Bug fix:
From: 	Juan Pablo Hierro Álvarez <BUSCAIDEAS@terra.es>
To: 	maxima@www.ma.utexas.edu
Subject: 	[Maxima] Installation instruction and a small bug.
Date: 	13 Apr 2002 09:58:49 +0200
First, I think there is a small bug in the file todd-coxeter.lisp
The patch would be
5c5
< (proclaim '(type (vector t)  $todd_coxeter_state))
---
> (proclaim '(type (vector (t))  $todd_coxeter_state))
21c21
<     (declare (fixnum nvars) (type (vector t) multiply-table))
---
>     (declare (fixnum nvars) (type (vector (t)) multiply-table))

With the original version (vector (t)), it complains that the type (t) is not
defined. But t is, of course, a type by itself.