Contributed file from Dan Stanger. He writes:
Attached is a first pass of Lindstedt code, which has the same calling
sequence as the commercial function, for inclusion in the maxima
project.  I gave it the LGPL, but I can change that to the GPL if
required.  I would appreciate any comments about my coding style, as
well as other comments.