Improved indentation (with some side effects, such as skipping over strings when necessary while doing searches).
New variable: `maxima-dont-reindent-some-lines'
If non-nil, don't reindent comments beginning with
"/*" on an otherwise blank line.

When completing (M-TAB) while in a string, use file-name completion.
Consistently use the same type of completion (choice of completions
appears in a separate buffer).

C-cC-e sends previous Maxima form (similar to C-xC-e), and
C-cC-s sends previous form and goes to the next form.

Maxima minor mode:
 C-c=e
   `maxima-minibuffer-on-determined-region'
 C-c=l
   `maxima-minibuffer-on-line'
 C-c=r
   `maxima-minibuffer-on-region'
 C-c=f
   `maxima-minibuffer-on-form'
 C-c=m
   `maxima-minibuffer'
 C-c=o
   `maxima-insert-last-output'
 C-c=t
   `maxima-insert-last-output-tex'

Regions sent to the Maxima process (both in Maxima mode and Maxima
minor mode) are highlighted (until the next action).

Info help uses emacs's info more, so specifying the maxima info file
isn't necessary.
C-cC-d and C-cC-h prompt for (with completion) a symbol, and will go
directly to the proper info page.  With an argument, use the symbol
under the point.
C-cC-a prompts for an expression, and presents a list of Maxima symbols
containing the expression.  With an argument, use the symbol under the
point.