Patch $EV to fix SF bug report [ 1234221 ] Segmentation fault on peculiar syntax.
Previously $EV assumed that evey argument after the first one
is either a symbol, a number, or a list.  The input
  (%i1) a + b, load ("empty.mac");
where empty.mac is an empty file, caused a Lisp error because load returns
the name of the file as a Lisp string (not a Maxima string).
This error is avoided by testing for a Lisp string and printing an error
message "Improper argument to ev" in that case.

Perhaps load should return a Maxima string, but, in any event, $EV should
not barf on an unexpected argument.

Patch tested on 5.9.1cvs, clisp 2.33.1 and gcl 2.6.6, on linux.
run_testsuite reports no unexpected errors.