Fix bugs in atanh.
src/comm2.lisp:
o Correct the workaround in simpatan2 for a buggy fpatan2.

src/float.lisp:
o fpatan2 was incorrectly computing the result for x > 0 and y < 0.
o Add comments on how complex-atanh works.
o Fix errors in implementation.  (This wasn't even close to the
  reference version in CMUCL!)

With these changes most of the failures in rtest_trig, problem 68 are
gone, except for atanh(+/-2).