Bug [ 1385307 ] 2*2^k doesn't simplify
This is an attempt to fix this bug.

o In TMS, if the factor is a number, we look at all the terms in the
  product to see if any are of the form factor^k.  If so, we modify
  that term with an appropriately adjusted exponent.

o Add EXPONENT-OF, which basically determines log(m) to the base b,
  but only if m is a power of b.

o TIMESIN reindented to make it easier to find the labels.

o Added two new cases to TIMESIN.
  - One case is number*a^k.  If number is some power of a, we convert
    that to a^(k+m) for an appropriate m.
  - Another case is a^k*rational.  This is handled as 2 separate
    operations: a^k*n and a^k/d, and each of these can be done as
    above.

There are still cases we don't handle like 2*3*2^k.  We might like the
answer to be 3*2^(k+1), but we actually return 6*2^k.  To handle that
case, we'd have to factor 6 and test each factor against the running
product.  I'm not going to add that.

Note that 3*2^k*2 is 3*2^(k+1), 2^k*3*2 is 3*2^(k+1) and 2*3*6^k is
6^(k+1).  I think this is because products are done left-to-right, in
order.