Bug [ 1620977 ] limit(5^n/(2^n*3^n),n,inf) is wrong
ISMAX incorrectly returns the strength of a product of terms.  It
returns the strength of the strongest term, but it should really be
the product of the terms.  That's why the limit was wrong.  ISMAX
thought the denominator grew like 3^n instead of 6^n.

src/limit.lisp:
o Fix ISMAX

tests/rtest15.mac:
o Add test