Bug [ 1635372 ] specint returns expression containing internal Lisp variable
Partial fix:  We return the noun form for the following cases:
fail-in-f2p105v2cond, fail-in-f50cond, and fail-in-dionarghyp-y.

Actually, the first two are now hidden and they will show up in
fail-in-dionarghyp-y.

These are hidden now, and actually
src/hypgeo.lisp:
o Return noun forms for the above cases

tests/rtest14.mac
o Add tests for the noun forms above.