Adding code to support the integration of special functions.
Short description of the modifications:

1. Put the integral of the elementary function on the property list.
2. Extend the routine integrallookups with an algorithm to look up
   the integral from the property list.
3. Write a more general routine to do a partial integration and call this
   routine in the routine intform.
4. Generalize diffdiv to support the integration of functions with more
   than one argument (e.g. for gamma_incomplete and expintegral_e).

In a first step the following function have the elementary integral on the property list:

gamma_incomplete

expintegral_e
expintegral_e1
expintegral_ei
expintegral_li
expintegral_si and expintegral_ci
expintegral_shi and expintegral_chi

erf, erfc and erfi

Because of the general algorithm of the new function partial-integration and diffdiv a lot of additional functions can be integrated.

Tested with CLISP 2.44 and GCL 2.69. There are no problems with the new code for the integration of special functions.