Improving the tests for using the series expansion or the expansion in continued fractions
The series expansion for gamma_incomplete(a,z)  converges for a pure imaginary argument z to wrong results. The expansion in continued fractions has to be used. For negative realpart(z) the complex plane is devided into two parts. In the region containing the negative real axis the series expansion is used. In the other part of the complex plane which contains the imaginary axis an expansion in continued fractions is used.

Further changes because of some wrong tests with rational numbers which causes Lisp errors.

Removing some test code.

Changes in rtest_gamma.mac:
Adding more test values and adjusting two tests for the erf function.

Tested with CLISP 2.44 and GCL 2.6.9.