Implementation of a general factor for the half-angle-transformation of the functions sin, cos, sinh and cosh.
The transformation simplifies to correct expressions for real and complex arguments of the trigonometrix and hyperbolic functions.

In rtest_trig.mac restriction of the test in Problem 33 to a postive interval (0,3). The results of this test are given for this interval.

Related bug: SF [505443] Halfangle limitation

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.