Correcting a bug when using the flag $numer.
The bug was introduced with revision 1.30. Maxima gets a Lisp error using an expression and the flag $numer e.g. gamma(%e),numer --> Lisp error.

The line ($numer (gammafloat (fpcofrat j))) is removed. The check for numerical evaluation is  combined with the tests for float or bigfloat numbers. This is done by the functions  float-numerical-eval-p, bigfloat-numerical-eval-p, ...

Tested with GCL 2.6.8 and CLISP 2.44