Adding the functions risplit-bessel-j-or-i and risplit-bessel-k-or-y
These functions handle special cases when we know that the result is pure real or pure imaginary.

The algorithm of risplit looks on the property list for a function which handle the realpart and imagpart of a function.

The default behavior of Maxima is to assume that a function is always real valued when the arguments are real. This is changed by the functions risplit-bessel-j-or-i and risplit-bessel-k-or-y. These functions try to return a correct expression for real arguments too.

Tested with CLISP 2.44 and GCL 2.6.8. No problems with the testsuite.