Revision of the function beta_incomplete
More careful handling of the domain of the Incomplete Beta function.
Special handling of a negative integer as the first argument a for float and bigfloat calculation.
Do not use the gamma function in the reflection formula, but beta(a,b). beta(a,b) checks carefully the arguments.

Adding a utility function integer-representation-p(x). This function checks if we have a float or bigfloat representation of an integer.