Extension to laplace to call $specint.
When laplace fails to find the laplace transform $specint is called. $specint can handle specail functions.

Correct a bug in the routine createname. Now the integration of a lot of integrands works again.

Changes:

1. Adding the routine check-call-to-$specint and call it in the routine laplace
2. Set the flags errcatch and errormsg before a call to $defint
3. Do not use gentemp in the function createname but intern the symbol.

Further change in desolve.lisp:

Initialize the variable lvar with a gensym and not nil. lvar contains the parameter of the laplace transformation. This parameter is passed to the $laplace and $specint. $specint does not work well with a value nil as a parameter.

Tested with GCL 2.6.8 and CLISP 2.44. No problems with the testsuite.