Set the variable lvar to a gensym and not nil.
$laplace has changed and calls $specint when it fails to find the laplace transformation.

Initialize the variable lvar with a gensym and not nil. lvar contains the parameter of the laplace transformation. This parameter is passed to the functions $laplace and $specint. $specint does not work well with a value nil as a parameter.

Tested with GCL 2.6.8 and CLISP 2.44. The testsuite has no problems.