Improving the handling of complex expressions for the functions abs, carg and rectform, ...:
Check with $csign for a complex expression and call cabs.
With this extension complex expression like abs(%e^z) with z declared to be complex give a correct  result.

Check for an expression with $conjugate more early. This simplification has to be done before we  call cabs.

Remove old code at 3 places.